backend/dztour/requirements.txt
amqp==5.3.1
asgiref==3.10.0
asttokens==3.0.1
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
coverage==7.12.0
cryptography==46.0.3
decorator==5.2.1
Django==5.2.8
django-allauth==65.13.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-ratelimit==4.1.0
django-storages==1.14.6
Django-Verify-Email==3.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
executing==2.2.1
factory_boy==3.3.3
Faker==38.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
sentry-sdk==2.45.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
structlog==25.5.0
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
